+++ src/JSSoft.Commands/CommandParameterArrayAttribute.cs
@@ -1,0 +1,11 @@
+// <copyright file="CommandParameterArrayAttribute.cs" company="JSSoft">
+//   Copyright (c) 2024 Jeesu Choi. All Rights Reserved.
+//   Licensed under the MIT License. See LICENSE.md in the project root for license information.
+// </copyright>
+
+namespace JSSoft.Commands;
+
+[AttributeUsage(AttributeTargets.Parameter)]
+public sealed class CommandParameterArrayAttribute : CommandParameterBaseAttribute
+{
+}