+++ src/JSSoft.Commands/CommandParameterAttribute.cs
@@ -1,0 +1,10 @@
+// <copyright file="CommandParameterAttribute.cs" company="JSSoft">
+//   Copyright (c) 2024 Jeesu Choi. All Rights Reserved.
+//   Licensed under the MIT License. See LICENSE.md in the project root for license information.
+// </copyright>
+
+namespace JSSoft.Commands;
+
+public sealed class CommandParameterAttribute : CommandParameterBaseAttribute
+{
+}