+++ CHANGES.md
@@ -12,12 +12,17 @@ To be Released.
   [[#19]]
 * Added the feature to parse custom types.  [[#20]]
 * Added Allows custom types to be used for method parameters  [[#22]]
+* Added sub command of command  [[#23]]
+* Added Added a variety of attributes that can be used for parameters.  [[#25]]
+* Improved the structure to allow ICustomCommandDescriptor to be applied.  [[#25]]
 
 [#13]: https://github.com/s2quake/commands/pull/13
 [#17]: https://github.com/s2quake/commands/pull/17
 [#19]: https://github.com/s2quake/commands/pull/19
 [#20]: https://github.com/s2quake/commands/pull/20
 [#22]: https://github.com/s2quake/commands/pull/22
+[#23]: https://github.com/s2quake/commands/pull/23
+[#25]: https://github.com/s2quake/commands/pull/25
 
 
 6.0.1